pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "jinja2>=3.1.6",
    "openai>=2.7.1",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
